change prices in db to int

+++ app/src/main/java/com/example/navsample/dto/filter/FilterProductList.kt
@@ -3,8 +3,8 @@ package com.example.navsample.dto.filter
 data class FilterProductList(
     var store: String = "",
     var category: String = "",
-    var lowerPrice: Double = -1.0,
-    var higherPrice: Double = -1.0,
+    var lowerPrice: Int = -1,
+    var higherPrice: Int = -1,
     var dateFrom: String = "",
     var dateTo: String = ""
 )
